Modernize the core Skipjack code a bit.

+++ CrappyCrypto/Skipjack.h
@@ -16,8 +16,8 @@ void rule_a(uint8_t* block, const uint8_t* key, uint16_t counter);
 void rule_b(uint8_t* block, const uint8_t* key, uint16_t counter);
 void rule_a_inverse(uint8_t* block, const uint8_t* key, uint16_t counter);
 void rule_b_inverse(uint8_t* block, const uint8_t* key, uint16_t counter);
-uint16_t g_permutation(uint16_t w_block, const uint8_t* key, int step);
-uint16_t g_permutation_inverse(uint16_t w_block, const uint8_t* key, int step);
+uint16_t g_permutation(uint16_t w_block, const uint8_t* key, uint16_t step);
+uint16_t g_permutation_inverse(uint16_t w_block, const uint8_t* key, uint16_t step);
 
 }
